Rename test utils folder to be more meaningful.
